+++ Step 3B-2. Build Azure Advanced S2S - Region 2.ps1
@@ -1,3 +1,4 @@
+#Requires -Modules Az
 $ErrorActionPreference = "Stop"
 
 #region Grab Configurations
@@ -279,9 +280,10 @@ Elseif( $null -eq $currentGwConnection)
     }
 }
 Else{
-    Write-Host ("Gateway is not connected. Attempting to update vyos router vpn settings to Azure's public IP [{0}]..." -f $azpip.IpAddress) -ForegroundColor Yellow
-    $Global:sharedPSKKey = Get-AzVirtualNetworkGatewayConnectionSharedKey -Name $AzureAdvConfigSiteB.VnetConnectionName -ResourceGroupName $AzureAdvConfigSiteB.ResourceGroupName
-    $VyOSConfig['ResetVPNConfigs'] = $true
+    Write-Host ("Gateway is not connected! ") -ForegroundColor Red
+    Write-Host ("Unable to reset router vpn settings because it may remove Region 1 settings. Manual interventions is required!") -ForegroundColor Yellow
+    $VyOSConfig['ResetVPNConfigs'] = $False
+    $RouterAutomationMode = $false
 }
 #endregion
 
@@ -360,7 +362,9 @@ save
 "@
 #endregion
 
-
+#Always output script
+$ScriptName = $LogfileName.replace('.log','.script')
+$VyOSFinal -split '\n' | %{$_ | Set-Content "$PSScriptRoot\Logs\$ScriptName"}
 
 If($RouterAutomationMode)
 {
@@ -460,12 +464,11 @@ If($RunManualSteps){
     Write-Host ("Local Router Prefix:      {0}" -f $VyOSConfig.LocalCIDRPrefix)
     Write-Host ("Local Router External:    {0}" -f $VyOSConfig.LocalCIDRPrefix)
     Write-host ("Home Public IP:           {0}" -f $HomePublicIP)
-    Write-Host "Be sure to follow a the configuration file 'VyOS_vpn_2site_bgp.md' in the VyOS_setup folder`n" -ForegroundColor Yellow
+    Write-Host "Be sure to follow a the configuration file: '$PSScriptRoot\Logs\$ScriptName'`n" -ForegroundColor Yellow
 
-    $VyOSFinal -split '\n' | %{$_ | Add-Content "$PSScriptRoot\Logs\vyoss2sregion1setup.txt"}
     #region Copy Paste Mode
     Write-Host "`nOpen ssh session for $($VyOSConfig.VMName):`n" -ForegroundColor Yellow
-    Write-Host "Copy script below line or from $PSScriptRoot\Logs\vyoss2sregion1setup.txt" -ForegroundColor Yellow
+    Write-Host "Copy script below line or from $PSScriptRoot\Logs\$ScriptName" -ForegroundColor Yellow
     Write-Host "--------------------------------------------------------" -ForegroundColor Yellow
     Write-Host $VyOSFinal -ForegroundColor Gray
     Write-Host "--------------------------------------------------------" -ForegroundColor Yellow