+++ Step 3B-3. Connect Azure Advanced S2S Regions.ps1
@@ -1,3 +1,4 @@
+#Requires -Modules Az
 # set error preference
 $ErrorActionPreference = "Stop"