+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change log for AzureSite2SiteVpnLab
 
 
+## 1.3.3 - November 27, 2021
+
+- Simplified Azure connection; removed Connect-AzureEnvironment function; fixed azure subscription selection
+- Added requires check for each script.
+- Forced vyos scripts output always even if automatic; named same as log
+- Fixed Vyos lan switch attachments; created more than needed when similar named networks existed
+- Changed resources to lowercase; easier readability in Azure
+- Fixed vyos reset function; disabled function for region 2
+
 ## 1.3.2 - November 26, 2021
 
 - Added vpn check after rerun; allows script to fix the connection is ran again